fix(harness): harden invariant registry contract
